requirements.txt
datasets==2.2.1
deepspeed==0.10.3
editdistance==0.6.2
einops==0.7.0
fastapi==0.104.1
fire==0.5.0
flash_attn==2.3.6
gradio==4.8.0
huggingface_hub==0.19.4
ipython==8.11.0
langchain==0.0.348
langchain_experimental==0.0.45
numpy==1.23.5
openai==0.27.8
packaging==23.2
pandas==2.0.2
peft==0.2.0
Pillow==9.4.0
Pillow==10.1.0
pycocoevalcap==1.2
ray==2.8.1
Requests==2.31.0
shortuuid==1.0.11
tenacity==8.2.2
torch==2.0.1
torch_xla==2.1.0
tqdm==4.62.3
transformers==4.30.2
uvicorn==0.24.0.post1
